video_analysis_ms/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
pymongo==4.10.1
mediapipe==0.10.20
opencv-python==4.10.0.84
numpy==1.26.4
python-dotenv==1.0.1
requests==2.32.3